backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
pydantic==2.5.0
aiohttp==3.9.1
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.8.5

apscheduler==3.10.4
jinja2==3.1.2

azure-cognitiveservices-speech==1.34.0